+++ src/client/config.ts
@@ -1,3 +1,5 @@
+import { Client } from "./client";
+
 /**
  * @see https://kafka.apache.org/documentation/#brokerconfigs
  */
@@ -18,4 +20,30 @@ export namespace BrokerConfigs {
      */
     export const AUTO_CREATE_TOPIC_ENABLE = 'auto.create.topics.enable';
 
+    export interface AutoCreateTopicResult {
+        type: "enabled" | "disabled" | "unknown";
+        error?: any;
+    }
+
+    export async function getAutoCreateTopicEnabled(client: Client): Promise<AutoCreateTopicResult> {
+        try {
+            const brokers = await client?.getBrokers();
+
+            if (brokers) {
+                for (let i = 0; i < brokers.length; i++) {
+                    const configs = await client?.getBrokerConfigs(brokers[i].id);
+                    const config = configs?.find(ce => ce.configName === BrokerConfigs.AUTO_CREATE_TOPIC_ENABLE);
+                    if (config) {
+                        const type = config.configValue === 'true' ? "enabled" : "disabled";
+                        return { type };
+                    }
+                }
+            }
+
+            return { type: "unknown" };
+        }
+        catch (error) {
+            return { type: "unknown", error };
+        }
+    }
 }